Update ignored keywords, add DRY_RUN flag for movie name parse
- Update logging statements

+++ src/plex/movie_name_parser.py
@@ -2,11 +2,14 @@ import re
 import os
 import shutil
 
-
 from plex.util import BLACKLIST, roman_char_fix, capitalize_title, remove_unwanted_chars, remove_whitespaces, BackgroundColors
+
 
 # Change this to point to the directory containing movies
 TARGET_DIR = '/path/to/movies/directory'
+
+# Change this to False when ready to rename the files/directories
+IS_DRY_RUN = False
 
 
 """
@@ -48,16 +51,16 @@ def clean_directory(path_name: str, dir_name: str) -> None:
     dir_name = roman_char_fix(dir_name)
 
     # Replace if any changes
-    if dir_name != original_dir_name:
+    if dir_name != original_dir_name and not IS_DRY_RUN:
         try:
             shutil.move(os.path.join(path_name, original_dir_name), os.path.join(path_name, dir_name))
+            pass
         except OSError as e:
             if "Destination path" and "already exists" in str(e):
                 print(f"{BackgroundColors.WARNING}Dir already exists{BackgroundColors.ENDC}")
                 handle_directory_exists_error(dir_name, original_dir_name, path_name)
         else:
-            print(f"{BackgroundColors.OKGREEN}Original dir name: {original_dir_name}{BackgroundColors.ENDC}")
-            print(f"{BackgroundColors.OKGREEN}Final dir name: {dir_name}{BackgroundColors.ENDC}")
+            print(f"{BackgroundColors.OKGREEN}Renamed dir: {original_dir_name} -> {dir_name}{BackgroundColors.ENDC}")
 
 
 def handle_directory_exists_error(dir_name, original_dir_name, path_name):
@@ -96,13 +99,6 @@ def clean_file(path_name: str, file_name: str) -> None:
     original_file_name = file_name
     file_ext = os.path.splitext(file_name)[1][1:]
 
-    # Remove .random extension files
-    if ".random" in file_name.lower():
-        full_file_path = os.path.join(path_name, file_name)
-        os.remove(full_file_path)
-        print(f"{BackgroundColors.OKBLUE}Removed .random file: {full_file_path}{BackgroundColors.ENDC}")
-        return
-
     remove_empty_files(path_name)
     file_name = capitalize_title(file_name)
     file_name = remove_unwanted_chars(file_name)
@@ -114,15 +110,10 @@ def clean_file(path_name: str, file_name: str) -> None:
     file_name = f'{file_name}.{file_ext}'
 
     # Rename the files
-    if file_name != original_file_name:
-        try:
-            shutil.move(os.path.join(path_name, original_file_name),
-                        os.path.join(path_name, file_name))
-        except Exception as exp:
-            print(f"{BackgroundColors.FAIL}[Error] {exp}{BackgroundColors.ENDC}")
-        else:
-            print(f'Original filename: {original_file_name}')
-            print(f'Updated filename: {file_name}')
+    if file_name != original_file_name and not IS_DRY_RUN:
+        shutil.move(os.path.join(path_name, original_file_name),
+                    os.path.join(path_name, file_name))
+        print(f"{BackgroundColors.OKBLUE}Renamed file: {original_file_name} ->  {file_name}{BackgroundColors.ENDC}")
 
 
 def remove_duplicate_file_ext(file_name: str, file_ext: str) -> str:
@@ -144,7 +135,7 @@ def remove_empty_files(path_name):
 def year_fix(text: str) -> str:
     """ Update year in the title if it exists """
     text = remove_unwanted_chars(text)
-    parsed_name = re.split(r'([12][90]\d{2})', text, re.IGNORECASE)
+    parsed_name = re.split(r'([12][90]\d{2})', text)
 
     if len(parsed_name) > 1:
         title, year, *_ = parsed_name
@@ -165,7 +156,12 @@ def blacklist_word_fix(text: str) -> str:
 
 def main():
     """ Main function """
+
+    print(f"{BackgroundColors.OKCYAN}Running script under DRY RUN. Turn this off by setting `IS_DRY_RUN` flag to {IS_DRY_RUN} under {TARGET_DIR} directory.{BackgroundColors.ENDC}")
+
+    # Meat and potatoes
     get_all_files(TARGET_DIR)
+
     print(f"{BackgroundColors.OKCYAN}Completed running the script.{BackgroundColors.ENDC}")
